Add files via upload
I simulate the Hodgkin-Huxley equations with an external current applied. I also carry out an anaylisis of the stability of the fixed points depending on the value of the parameter for the external current I.

The hodgkin-huxlety equations simulate the membrante potential of a nerve cell and the conductances for different ions of this membrane. In this case an external current is also applied to stimulate the membrane. In my thesis I studied how the behaviour of the potential changes as we vary the value of the external current applied. The general behaviour is the well known action potential that is exhibited by the system if the external current is high enough to stimulate it.  If the current is held constant the action potential occurs one after another. However I found that for some values of the current the system reachs the equilibrium after some realizations of the action potential. I made a function to measure that time.

In my thesis I made further analysis than those shown here. I will try to upload more of the code I used as soon as  I organize and edit it.